Separate out tests for INSERT INTO
