+++ test/n1ql.test.ts
@@ -19,6 +19,7 @@ import supportsParams from './options/param';
 import supportsWindow from './features/window';
 import supportsSetOperations from './features/setOperations';
 import supportsLimiting from './features/limiting';
+import supportsInsertInto from './features/insertInto';
 
 describe('N1qlFormatter', () => {
   const language = 'n1ql';
@@ -47,6 +48,7 @@ describe('N1qlFormatter', () => {
   supportsParams(format, { positional: true, numbered: ['$'], named: ['$'] });
   supportsWindow(format);
   supportsLimiting(format, { limit: true, offset: true });
+  supportsInsertInto(format);
 
   it('formats INSERT with {} object literal', () => {
     const result = format(
